Refactor: type migrate entity ids from int -> long.

+++ src/main/java/dk/os2opgavefordeler/model/DistributionRule.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 public class DistributionRule implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 
 	@ManyToOne
 	@JoinColumn
@@ -47,7 +47,7 @@ public class DistributionRule implements Serializable {
 
 //	@ManyToOne
 //	private Employment assignedEmp;
-	private int assignedEmp;
+	private long assignedEmp;
 
 	public Optional<OrgUnit> getResponsibleOrg() {
 		return Optional.ofNullable(responsibleOrg);
@@ -65,11 +65,11 @@ public class DistributionRule implements Serializable {
 		this.assignedOrg = assignedOrg;
 	}
 
-	public int getAssignedEmp() {
+	public long getAssignedEmp() {
 		return assignedEmp;
 	}
 
-	public void setAssignedEmp(int assignedEmp) {
+	public void setAssignedEmp(long assignedEmp) {
 		this.assignedEmp = assignedEmp;
 	}
 //	@OneToMany
@@ -93,7 +93,7 @@ public class DistributionRule implements Serializable {
 		return Optional.ofNullable(parent);
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
