+++ src/main/java/dk/os2opgavefordeler/model/presentation/DistributionRulePO.java
@@ -14,12 +14,12 @@ public class DistributionRulePO {
 	/**
 	 * Id of the topic route for identification.
 	 */
-	private int id;
+	private long id;
 
 	/**
 	 * Specifies parent KLE id.
 	 */
-	private int parent;
+	private long parent;
 
 	/**
 	 * The KLE data associated with this.
@@ -29,44 +29,44 @@ public class DistributionRulePO {
 	/**
 	 * The organisational unit responsible for handling this topic.
 	 */
-	private int org;
+	private long org;
 
 	/**
 	 * Employee responsible for handling this topic.
 	 */
-	private int employee;
+	private long employee;
 
 	/**
 	 * OrgUnit with responsibility for (ownership of) the topic, who can decide who handles the topic.
 	 */
-	private int responsible;
+	private long responsible;
 
 	public DistributionRulePO() {
 	}
 
 	public DistributionRulePO(DistributionRule source) {
-		this.parent = source.getParent().map(rule -> rule.getId()).orElse(0);
+		this.parent = source.getParent().map(rule -> rule.getId()).orElse(0L);
 		this.id = source.getId();
 
 		this.kle = kleFrom(source.getKle());
-		this.responsible = source.getResponsibleOrg().map(OrgUnit::getId).orElse(0);
+		this.responsible = source.getResponsibleOrg().map(OrgUnit::getId).orElse(0L);
 		this.employee = source.getAssignedEmp();
-		this.org = source.getAssignedOrg().map(OrgUnit::getId).orElse(0);
+		this.org = source.getAssignedOrg().map(OrgUnit::getId).orElse(0L);
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getParent() {
+	public long getParent() {
 		return parent;
 	}
 
-	public void setParent(int parent) {
+	public void setParent(long parent) {
 		this.parent = parent;
 	}
 
@@ -78,27 +78,27 @@ public class DistributionRulePO {
 		this.kle = kle;
 	}
 
-	public int getOrg() {
+	public long getOrg() {
 		return org;
 	}
 
-	public void setOrg(int org) {
+	public void setOrg(long org) {
 		this.org = org;
 	}
 
-	public int getEmployee() {
+	public long getEmployee() {
 		return employee;
 	}
 
-	public void setEmployee(int employee) {
+	public void setEmployee(long employee) {
 		this.employee = employee;
 	}
 
-	public int getResponsible() {
+	public long getResponsible() {
 		return responsible;
 	}
 
-	public void setResponsible(int responsible) {
+	public void setResponsible(long responsible) {
 		this.responsible = responsible;
 	}
